// convex/types.ts
export type UniFiWebhookBody = {
  app: string;
  deviceEventClassId: string;
  message: string;
  name: string;
  parameters: {
    proto: string;
    src: string;
    spt: string;
    dst: string;
    dpt: string;
    act: string;
    app: string;
    UNIFIcategory: string;
    UNIFIsubCategory: string;
    UNIFIhost: string;
    UNIFIrisk: string;
    UNIFIpolicyName: string;
    UNIFIpolicyType: string;
    UNIFIdirection: string;
    deviceOutboundInterface: string;
    UNIFIdeviceMac: string;
    UNIFIdeviceName: string;
    UNIFIdeviceModel: string;
    UNIFIdeviceIp: string;
    UNIFIdeviceVersion: string;
    UNIFIsrcRegion: string;
    UNIFIdstZone: string;
    UNIFItotalBytes: string;
    UNIFItotalPackets: string;
    UNIFIpacketsReceived: string;
    UNIFIpacketsSent: string;
    UNIFIbytesReceived: string;
    UNIFIbytesSent: string;
    UNIFIflowCount: string;
    UNIFIflowId: string;
    UNIFIflowStartTime: string;
    UNIFIipsSessionId: string;
    UNIFIipsSignature: string;
    UNIFIipsSignatureId: string;
    UNIFIutcTime: string;
  };
  severity: number;
  version: string;
  alarm_id: string;
};